refactor: replace response Log type with a new Log type and update imports across components
